+++ packages/keyring-eth-trezor/src/index.ts
@@ -2,5 +2,7 @@ export * from './trezor-keyring';
 export * from './trezor-keyring-v2';
 export * from './onekey-keyring';
 export * from './onekey-keyring-v2';
+export * from './trezor-error-handler';
+export * from './trezor-errors';
 export type * from './trezor-bridge';
 export * from './trezor-connect-bridge';